Add FABRIK constraint
This constraint is really just demonstrated as a distance constraint
applied in both directions. It stands for "Forwards and Backwards
Reaching Inverse Kinematics". It's an iterative solution for solving the
inverse kinematics problem.

This also comes with a bit of a refactor on how to construct the bodies
to support generating them from a specification, since some of the
implementations depend inner types and boxed traits types.